WIP sdl3 port

+++ Source/relive_lib/Sound/Reverb.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "SoundSDL.hpp"
-#include "SDL.h"
+#include "SDL3/SDL.h"
 
 #if USE_SDL2_SOUND
 
@@ -9,6 +9,6 @@ void Reverb_Init(s32 sampleRate);
 void Reverb_DeInit();
 void Reverb_PushSample(StereoSample_S16 v);
 void Reverb_Update(s32 index);
-void Reverb_Mix(StereoSample_S16* dst, SDL_AudioFormat format, Uint32 len, s32 volume);
+void Reverb_Mix(StereoSample_S16* dst, SDL_AudioFormat format, Uint32 len, f32 volume);
 
 #endif
